feat(YB-168): TeamUtil 을 제거하고 RedisTeamService 에서 팀정보를 받아올 수 있도록 함
